apps/synthesizer/requirements.txt
upstash-redis
pymongo[srv]
python-dotenv
scikit-learn
numpy
google-generativeai
certifi